refactor: break out queryFn to own component
